Additions/VBoxServices: two warnings
svn:sync-xref-src-repo-rev: r84115

+++ src/VBox/Additions/common/VBoxService/VBoxServiceControlSession.cpp
@@ -1,4 +1,4 @@
-/* $Id: VBoxServiceControlSession.cpp 44869 2013-02-28 15:42:02Z andreas.loeffler@oracle.com $ */
+/* $Id: VBoxServiceControlSession.cpp 44926 2013-03-05 15:04:26Z noreply@oracle.com $ */
 /** @file
  * VBoxServiceControlSession - Guest session handling. Also handles
  *                             the forked session processes.
@@ -246,7 +246,7 @@ static int gstcntlSessionHandleFileRead(uint32_t uClientId, uint32_t cParms,
             rc = VERR_NOT_FOUND;
 
         /* Report back in any case. */
-        CALLBACKPAYLOAD_FILE_NOTFIY_READ cplRead = { rc, cbRead, pvDataRead };
+        CALLBACKPAYLOAD_FILE_NOTFIY_READ cplRead = { rc, (uint32_t)cbRead, pvDataRead };
         int rc2 = VbglR3GuestCtrlFileNotify(uClientId, ctx.uContextID, GUEST_FILE_NOTIFYTYPE_READ,
                                             &cplRead, sizeof(cplRead));
         if (   cbToRead > cbScratchBuf
@@ -297,7 +297,7 @@ static int gstcntlSessionHandleFileReadAt(uint32_t uClientId, uint32_t cParms,
             rc = VERR_NOT_FOUND;
 
         /* Report back in any case. */
-        CALLBACKPAYLOAD_FILE_NOTFIY_READ cplRead = { rc, cbRead, pvDataRead };
+        CALLBACKPAYLOAD_FILE_NOTFIY_READ cplRead = { rc, (uint32_t)cbRead, pvDataRead };
         int rc2 = VbglR3GuestCtrlFileNotify(uClientId, ctx.uContextID, GUEST_FILE_NOTIFYTYPE_READ,
                                             &cplRead, sizeof(cplRead));
         if (   cbToRead > cbScratchBuf
